[volume-3] 도메인모델 및 서비스 구현

+++ core/service/user-service/src/main/java/com/loopers/core/service/user/command/UserPointChargeCommand.java
@@ -2,13 +2,15 @@ package com.loopers.core.service.user.command;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class UserPointChargeCommand {
 
     private final String userIdentifier;
-    private final int point;
+    private final BigDecimal point;
 
-    public UserPointChargeCommand(String userIdentifier, int point) {
+    public UserPointChargeCommand(String userIdentifier, BigDecimal point) {
         this.userIdentifier = userIdentifier;
         this.point = point;
     }
